Added delete to mailbox

+++ src/main/java/edu/stevens/ssw690/DuckSource/dao/MailboxDao.java
@@ -13,7 +13,7 @@ public interface MailboxDao {
 	public List<Mailbox> getByUserId(Integer id);
 	public List<Mailbox> getUserInbox(Integer id);
 	public List<Mailbox> getUserSent(Integer id);
-	public Mailbox findById(BigInteger id);
+	public Mailbox findById(Integer id);
 	public Mailbox getByMessageId(Integer id);
 	public void remove(Mailbox mailbox);
 	public int getUnreadCount(Integer id);
