+++ src/main/java/edu/stevens/ssw690/DuckSource/dao/MailboxDaoImpl.java
@@ -84,7 +84,7 @@ public class MailboxDaoImpl implements MailboxDao {
 	        em.flush();
 	}
 	
-	public Mailbox findById(BigInteger id) {
+	public Mailbox findById(Integer id) {
     	return em.find(Mailbox.class, id);
 	}
 }
